backend/requirements.txt
# FastAPI & Web
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0

# Data validation
pydantic==2.7.4
pydantic-settings==2.3.3

# Image processing
Pillow==12.1.1
opencv-python-headless==4.10.0.82
numpy==1.26.4

# DICOM parsing
pydicom==2.4.4

# MRI processing (medpy requires SimpleITK)
SimpleITK==2.3.1

# Deep learning (CPU-only for portability; replace torch with GPU build as needed)
# torch>=2.6.0 fixes CVE for torch.load weights_only bypass (< 2.6.0)
torch==2.6.0
torchvision==0.21.0

# Report generation
python-docx==1.1.2

# HTTP requests (for external NLG API)
aiohttp==3.13.3
tenacity==8.3.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
